grids, ants, and pheromones
set up some empty functions in ant.cpp, but mostly worked on ant tiles and grid system. grid now spawns tiles in correct positions. need to implement more to see if tiles correctly manage their pheromones. considering refactoring so that only the grid is an actor that ticks, but /shrug